+++ docs/changelog.md
@@ -11,6 +11,7 @@
     + "History" page has been merged in this page.
 + Fix a solver bug of the stop button on inputs page.
 + Separate out "preferences" page as dialog.
++ Add "open actions" settings for the opening macros.
 
 **Development**: